+++ lib/rest/Voice.d.ts
@@ -8,7 +8,11 @@
 import Domain = require('../base/Domain');
 import Twilio = require('./Twilio');
 import V1 = require('./voice/V1');
+import { ByocTrunkListInstance } from './voice/v1/byocTrunk';
+import { ConnectionPolicyListInstance } from './voice/v1/connectionPolicy';
 import { DialingPermissionsListInstance } from './voice/v1/dialingPermissions';
+import { IpRecordListInstance } from './voice/v1/ipRecord';
+import { SourceIpMappingListInstance } from './voice/v1/sourceIpMapping';
 
 
 declare class Voice extends Domain {
@@ -19,7 +23,11 @@ declare class Voice extends Domain {
    */
   constructor(twilio: Twilio);
 
+  readonly byocTrunks: ByocTrunkListInstance;
+  readonly connectionPolicies: ConnectionPolicyListInstance;
   readonly dialingPermissions: DialingPermissionsListInstance;
+  readonly ipRecords: IpRecordListInstance;
+  readonly sourceIpMappings: SourceIpMappingListInstance;
   readonly v1: V1;
 }