+++ lib/rest/supersim/V1.d.ts
@@ -11,6 +11,8 @@ import { CommandList } from './v1/command';
 import { CommandListInstance } from './v1/command';
 import { FleetList } from './v1/fleet';
 import { FleetListInstance } from './v1/fleet';
+import { NetworkAccessProfileList } from './v1/networkAccessProfile';
+import { NetworkAccessProfileListInstance } from './v1/networkAccessProfile';
 import { SimList } from './v1/sim';
 import { SimListInstance } from './v1/sim';
 import { UsageRecordList } from './v1/usageRecord';
@@ -27,6 +29,7 @@ declare class V1 extends Version {
 
   readonly commands: CommandListInstance;
   readonly fleets: FleetListInstance;
+  readonly networkAccessProfiles: NetworkAccessProfileListInstance;
   readonly sims: SimListInstance;
   readonly usageRecords: UsageRecordListInstance;
 }