+++ spec/integration/rest/api/v2010/account/sip/domain.spec.js
@@ -55,7 +55,9 @@ describe('Domain', function() {
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
                   'emergency_calling_enabled': true,
-                  'secure': true
+                  'secure': true,
+                  'byoc_trunk_sid': 'BYaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'emergency_caller_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],
           'start': 0,
@@ -99,7 +101,9 @@ describe('Domain', function() {
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
                   'emergency_calling_enabled': true,
-                  'secure': true
+                  'secure': true,
+                  'byoc_trunk_sid': 'BYaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'emergency_caller_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],
           'start': 0,
@@ -148,7 +152,9 @@ describe('Domain', function() {
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
                   'emergency_calling_enabled': true,
-                  'secure': true
+                  'secure': true,
+                  'byoc_trunk_sid': 'BYaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'emergency_caller_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],
           'start': 0,
@@ -215,7 +221,9 @@ describe('Domain', function() {
                   'voice_status_callback_url': null,
                   'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
                   'emergency_calling_enabled': true,
-                  'secure': true
+                  'secure': true,
+                  'byoc_trunk_sid': 'BYaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'emergency_caller_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
           ],
           'start': 0,
@@ -318,7 +326,9 @@ describe('Domain', function() {
           'voice_status_callback_url': null,
           'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
           'emergency_calling_enabled': true,
-          'secure': true
+          'secure': true,
+          'byoc_trunk_sid': 'BYaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'emergency_caller_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       };
 
       holodeck.mock(new Response(201, body));
@@ -383,7 +393,9 @@ describe('Domain', function() {
           'voice_status_callback_url': null,
           'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
           'emergency_calling_enabled': true,
-          'secure': true
+          'secure': true,
+          'byoc_trunk_sid': 'BYaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'emergency_caller_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       };
 
       holodeck.mock(new Response(200, body));
@@ -447,7 +459,9 @@ describe('Domain', function() {
           'voice_status_callback_url': null,
           'voice_url': 'https://dundermifflin.example.com/twilio/app.php',
           'emergency_calling_enabled': true,
-          'secure': true
+          'secure': true,
+          'byoc_trunk_sid': 'BYaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'emergency_caller_sid': 'PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       };
 
       holodeck.mock(new Response(200, body));