+++ spec/integration/rest/supersim/v1/fleet.spec.js
@@ -61,6 +61,7 @@ describe('Fleet', function() {
           'commands_enabled': true,
           'commands_method': 'GET',
           'commands_url': 'https://google.com',
+          'network_access_profile_sid': 'HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'sid': 'HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       };
@@ -109,6 +110,7 @@ describe('Fleet', function() {
           'commands_enabled': true,
           'commands_method': 'POST',
           'commands_url': null,
+          'network_access_profile_sid': 'HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'sid': 'HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       };
@@ -128,13 +130,13 @@ describe('Fleet', function() {
     function(done) {
       var body = {
           'meta': {
-              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0',
+              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0',
               'key': 'fleets',
               'next_page_url': null,
               'page': 0,
               'page_size': 50,
               'previous_page_url': null,
-              'url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0'
+              'url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0'
           },
           'fleets': [
               {
@@ -147,6 +149,7 @@ describe('Fleet', function() {
                   'commands_enabled': true,
                   'commands_method': 'POST',
                   'commands_url': null,
+                  'network_access_profile_sid': 'HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'sid': 'HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
@@ -160,13 +163,13 @@ describe('Fleet', function() {
     function(done) {
       var body = {
           'meta': {
-              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0',
+              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0',
               'key': 'fleets',
               'next_page_url': null,
               'page': 0,
               'page_size': 50,
               'previous_page_url': null,
-              'url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0'
+              'url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0'
           },
           'fleets': [
               {
@@ -179,6 +182,7 @@ describe('Fleet', function() {
                   'commands_enabled': true,
                   'commands_method': 'POST',
                   'commands_url': null,
+                  'network_access_profile_sid': 'HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'sid': 'HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
@@ -197,13 +201,13 @@ describe('Fleet', function() {
     function(done) {
       var body = {
           'meta': {
-              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0',
+              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0',
               'key': 'fleets',
               'next_page_url': null,
               'page': 0,
               'page_size': 50,
               'previous_page_url': null,
-              'url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0'
+              'url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0'
           },
           'fleets': [
               {
@@ -216,6 +220,7 @@ describe('Fleet', function() {
                   'commands_enabled': true,
                   'commands_method': 'POST',
                   'commands_url': null,
+                  'network_access_profile_sid': 'HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'sid': 'HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
@@ -250,13 +255,13 @@ describe('Fleet', function() {
       var body = {
           'fleets': [],
           'meta': {
-              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0',
+              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0',
               'key': 'fleets',
               'next_page_url': null,
               'page': 0,
               'page_size': 50,
               'previous_page_url': null,
-              'url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0'
+              'url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0'
           }
       };
 
@@ -275,13 +280,13 @@ describe('Fleet', function() {
     function(done) {
       var body = {
           'meta': {
-              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0',
+              'first_page_url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0',
               'key': 'fleets',
               'next_page_url': null,
               'page': 0,
               'page_size': 50,
               'previous_page_url': null,
-              'url': 'https://supersim.twilio.com/v1/Fleets?PageSize=50&Page=0'
+              'url': 'https://supersim.twilio.com/v1/Fleets?NetworkAccessProfile=HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&PageSize=50&Page=0'
           },
           'fleets': [
               {
@@ -294,6 +299,7 @@ describe('Fleet', function() {
                   'commands_enabled': true,
                   'commands_method': 'POST',
                   'commands_url': null,
+                  'network_access_profile_sid': 'HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'sid': 'HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
               }
@@ -344,6 +350,7 @@ describe('Fleet', function() {
           'commands_enabled': true,
           'commands_method': 'POST',
           'commands_url': null,
+          'network_access_profile_sid': 'HAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'sid': 'HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://supersim.twilio.com/v1/Fleets/HFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
       };